BySideScheme_backend/requirements.txt
openai
mem0ai
python-dotenv
pydantic
fastapi
uvicorn
pyautogen==0.2.35
PyYAML
